Add GrowableListWidgetFactory
