+++ engine/src/main/java/org/terasology/rendering/nui/widgets/types/builtin/CollectionWidgetFactory.java
@@ -16,25 +16,16 @@
 package org.terasology.rendering.nui.widgets.types.builtin;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMultiset;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedSet;
-import com.google.common.collect.Multiset;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.terasology.reflection.TypeInfo;
 import org.terasology.reflection.reflect.ConstructorLibrary;
 import org.terasology.reflection.reflect.ObjectConstructor;
 import org.terasology.rendering.nui.UIWidget;
 import org.terasology.rendering.nui.databinding.Binding;
-import org.terasology.rendering.nui.layouts.ColumnLayout;
-import org.terasology.rendering.nui.layouts.MultiRowLayout;
-import org.terasology.rendering.nui.layouts.RowLayout;
-import org.terasology.rendering.nui.layouts.RowLayoutHint;
-import org.terasology.rendering.nui.widgets.UIButton;
 import org.terasology.rendering.nui.widgets.types.TypeWidgetFactory;
 import org.terasology.rendering.nui.widgets.types.TypeWidgetLibrary;
-import org.terasology.utilities.ReflectionUtil;
+import org.terasology.rendering.nui.widgets.types.builtin.util.GrowableListWidgetFactory;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -44,8 +35,6 @@ import java.util.Set;
 import java.util.SortedSet;
 
 public class CollectionWidgetFactory implements TypeWidgetFactory {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CollectionWidgetFactory.class);
-
     private ConstructorLibrary constructorLibrary = new ConstructorLibrary();
 
     @Override
@@ -63,153 +52,57 @@ public class CollectionWidgetFactory implements TypeWidgetFactory {
             binding.set(constructor.construct());
         }
 
-        return Optional.of(createWidget((Binding<Collection<Object>>) binding, (TypeInfo<Collection<Object>>) type, library));
+        UIWidget widget = new GrowableListCollectionWidgetFactory<>(
+            (Binding<Collection<Object>>) binding,
+            (TypeInfo<Collection<Object>>) type,
+            library
+        )
+                              .create();
+
+        return Optional.of(widget);
     }
 
-    private <T extends Collection<E>, E> ColumnLayout createWidget(
-        Binding<T> binding,
-        TypeInfo<T> type,
-        TypeWidgetLibrary library
-    ) {
-        @SuppressWarnings({"unchecked"})
-        TypeInfo<E> elementType = (TypeInfo<E>) TypeInfo.of(
-            ReflectionUtil.getTypeParameterForSuper(type.getType(), Collection.class, 0)
-        );
+    private static class GrowableListCollectionWidgetFactory<T extends Collection<E>, E>
+        extends GrowableListWidgetFactory<T, E> {
+        public GrowableListCollectionWidgetFactory(Binding<T> binding, TypeInfo<T> type, TypeWidgetLibrary library) {super(binding, type, library);}
 
-        List<E> elementList = new ArrayList<>(binding.get());
-
-        ColumnLayout mainLayout = new ColumnLayout();
-
-        mainLayout.setFillVerticalSpace(false);
-        mainLayout.setAutoSizeColumns(false);
-
-        populateCollectionLayout(binding, type, library, elementType, elementList, mainLayout);
-
-        return mainLayout;
-    }
-
-    private <T extends Collection<E>, E> void populateCollectionLayout(
-        Binding<T> binding,
-        TypeInfo<T> type,
-        TypeWidgetLibrary library,
-        TypeInfo<E> elementType,
-        List<E> elementList,
-        ColumnLayout collectionLayout
-    ) {
-        UIButton addElementButton = new UIButton();
-
-        // TODO: Translate
-        addElementButton.setText("Add Element");
-        addElementButton.subscribe(widget -> {
-            elementList.add(null);
-
-            Optional<UIWidget> elementLayout = createElementLayout(
-                binding, type, library,
-                elementType, elementList, elementList.size() - 1, collectionLayout
-            );
-
-            if (!elementLayout.isPresent()) {
-                return;
+        private void updateBindingWith(Binding<T> binding, TypeInfo<T> type, Collection<E> items) {
+            try {
+                binding.get().clear();
+                binding.get().addAll(items);
+            } catch (UnsupportedOperationException e) {
+                // Bound collection is unmodifiable, create new
+                binding.set(newImmutableCollection(type, items));
             }
-
-            collectionLayout.addWidget(elementLayout.get());
-        });
-
-        collectionLayout.addWidget(addElementButton);
-
-        for (int i = 0; i < elementList.size(); i++) {
-            Optional<UIWidget> elementLayout = createElementLayout(
-                binding, type, library,
-                elementType, elementList, i, collectionLayout
-            );
-
-            if (!elementLayout.isPresent()) {
-                continue;
-            }
-
-            collectionLayout.addWidget(elementLayout.get());
-        }
-    }
-
-    private <T extends Collection<E>, E> Optional<UIWidget> createElementLayout(
-        Binding<T> binding,
-        TypeInfo<T> type,
-        TypeWidgetLibrary library,
-        TypeInfo<E> elementType,
-        List<E> elementList,
-        int elementIndex,
-        ColumnLayout collectionLayout
-    ) {
-        Optional<UIWidget> elementWidget = library.getWidget(
-            new Binding<E>() {
-                @Override
-                public E get() {
-                    return elementList.get(elementIndex);
-                }
-
-                @Override
-                public void set(E value) {
-                    elementList.set(elementIndex, value);
-                    updateBindingWith(binding, type, elementList);
-                }
-            },
-            elementType
-        );
-
-        if (!elementWidget.isPresent()) {
-            LOGGER.error(
-                "Could not get widget for element {} in collection",
-                elementList.get(elementIndex)
-            );
-            return Optional.empty();
         }
 
-        UIButton removeButton = new UIButton();
-        // TODO: Translate
-        removeButton.setText("-");
+        private T newImmutableCollection(TypeInfo<T> type, Collection<E> items) {
+            Class<T> rawType = type.getRawType();
 
-        removeButton.subscribe(widget -> {
-            elementList.remove(elementIndex);
+            // If the bound collection is unmodifiable, it must either be a standard
+            // Collection or a guava ImmutableCollection, so casts always succeed
+
+            // TODO: Support more Guava types?
+
+            if (SortedSet.class.isAssignableFrom(rawType)) {
+                return (T) ImmutableSortedSet.copyOf(items);
+            }
+
+            if (Set.class.isAssignableFrom(rawType)) {
+                return (T) ImmutableSet.copyOf(items);
+            }
+
+            return (T) ImmutableList.copyOf(items);
+        }
+
+        @Override
+        protected void updateBindingElements(List<E> elementList) {
             updateBindingWith(binding, type, elementList);
-
-            collectionLayout.removeAllWidgets();
-            populateCollectionLayout(binding, type, library, elementType, elementList, collectionLayout);
-        });
-
-        RowLayout elementLayout = new RowLayout();
-
-        elementLayout.addWidget(removeButton, new RowLayoutHint().setUseContentWidth(true));
-        elementLayout.addWidget(elementWidget.get(), new RowLayoutHint().setUseContentWidth(false));
-
-        return Optional.of(elementLayout);
-    }
-
-    private <T extends Collection<E>, E> void updateBindingWith(Binding<T> binding, TypeInfo<T> type, Collection<E> items) {
-        try {
-            binding.get().clear();
-            binding.get().addAll(items);
-        } catch (UnsupportedOperationException e) {
-            // Bound collection is unmodifiable, create new
-            binding.set(newImmutableCollection(type, items));
-        }
-    }
-
-    private <T extends Collection<E>, E> T newImmutableCollection(TypeInfo<T> type, Collection<E> items) {
-        Class<T> rawType = type.getRawType();
-
-        // If the bound collection is unmodifiable, it must either be a standard
-        // Collection or a guava ImmutableCollection, so casts always succeed
-
-        // TODO: Support more Guava types?
-
-        if (SortedSet.class.isAssignableFrom(rawType)) {
-            return (T) ImmutableSortedSet.copyOf(items);
         }
 
-        if (Set.class.isAssignableFrom(rawType)) {
-            return (T) ImmutableSet.copyOf(items);
+        @Override
+        protected List<E> getBindingCopy() {
+            return new ArrayList<>(binding.get());
         }
-
-        return (T) ImmutableList.copyOf(items);
     }
 }
